Some review fixes
